feat(ui): add mobile splash video and webp auto-upload
- Introduce `MobileSplashVideo` for the storefront entry.
- Implement automatic WebP conversion for banner image uploads.
- Update admin config schema and UI to support advanced storefront customization.
- Adjust navigation visibility for improved mobile web app responsiveness.

+++ src/data/mockSettings.ts
@@ -91,7 +91,7 @@ export const mockWebsiteSettings: WebsiteSettings = {
   categoryShowcaseEyebrowTag: "CURATED SELECTIONS",
   categoryShowcaseTitle: "Shop by Category Showcase",
   categoryShowcaseSubtitle: "Explore our spectacular hand-crafted designs categorized for perfect visual navigation",
-  categoryShowcaseLayout: "single",
+  categoryShowcaseLayout: "double",
   categoryShowcaseHeaderBgColor: "transparent",
   categoryShowcaseHeaderTextColor: "#052e16",
   categoryShowcaseHeaderBorderColor: "transparent",
